Updated disableTEXA.ps1
Updated disable script with 4 lines of code to kill processes related to TEXA. This helps Techs with another troubleshoot to close TEXA completely when TEXA is stuck loading something or stuck on a black screen and cannot close it. This way techs do not need to open task manager and risk terminating an important process.